Improve handing of id param in BP_Invitation::get().
`BP_Invitation::get_where_sql()` used loose logic when processing the `id` param. This changeset tightens up the logic to avoid database errors when faced with empty arrays or strings and ensures that no invitations are returned for "zero-like" values. Here are the expected responses for certain `id` params:
• `id = array()` should return no results.
• `id = array( 0 )` should return no results.
• `id = 0` should return no results.
• `id = "0"` should return no results.
• `id = ""` should return no results.
• `id = false` should result in no `id IN ()` statement being added to the parameter.
• `id = null` should result in no `id IN ()` statement being added to the parameter.

See #8149.